+++ src/TLE94112-ino.hpp
@@ -32,11 +32,34 @@
  *
  * @brief represents the TLE94112 base class
  *
- * This file exists only for compatibility with the Arduino setup.
- * 
+ * This class provides a simple API for connecting the TLE94112 shield via SPI.
+ * The 12 halfbridges, the PWM units and the registers are set and controlled
+ * from the library.
+ *
  */
 
-/** @} */
+class Tle94112Ino: virtual public TLE94112Ino
+{
+	public:
+		//! \brief standard constructor with default pin assignment
+		Tle94112Ino(void);
+
+		/*!
+		 * \brief constructor with individual pin assignment
+		 *
+		 * \param bus    void pointer to the object representing the SPI class
+		 * \param csPin  pin number of the CS pin
+		 */
+		Tle94112Ino(SPIClass &bus, uint8_t csPin);
+
+		//! \brief enables and initializes the TLE94112
+		void begin(void);
+
+		//! \brief deactivates all outputs and disables the TLE94112
+		void end(void);
+
+};
+/* @} */
 
 #endif /** TLE94112_FRAMEWORK **/
 #endif /** TLE94112_INO_HPP_ **/
