+++ src/config/tle94112-conf-dfl.hpp
@@ -18,7 +18,6 @@
 #ifndef TLE94112_FRAMEWORK
 #error  'TLE94112 error framework undefined'    /**< TLE94112 framework must be defined in conf. or at compilation */
 #endif
-
 /** @} */
 
 #endif /** TLE94112_CONF_DFL_HPP_ **/
