+++ src/config/tle94112-conf-opts.hpp
@@ -15,12 +15,8 @@
 #ifndef TLE94112_CONF_OPTS_HPP_
 #define TLE94112_CONF_OPTS_HPP_
 
-#define TLE94112_FRMWK_ARDUINO   0x01U
-#define TLE94112_FRMWK_WICED     0x02U
-#define TLE94112_FRMWK_RPI       0x03U
-#define TLE94112_FRMWK_MTB       0x04U
-#define TLE94112_FRMWK_PSOCC     0x05U
-
+#define TLE94112_FRMWK_ARDUINO   0x00U
+#define TLE94112_FRMWK_WICED     0x01U
 /** @} */
 
 #endif /** TLE94112_CONF_OPTS_HPP_ **/
