+++ src/framework/arduino/pal/gpio-arduino.hpp
@@ -30,7 +30,6 @@
 class GPIOIno: virtual public GPIO
 {
 private:
-	#define     UNUSED_PIN    0xFF  /**< Unused pin */
 	uint8_t     pin;
 	uint8_t     mode;
 	VLogic_t    logic;
