+++ src/framework/arduino/wrapper/TLE94112Motor-platf-ino.hpp
@@ -19,19 +19,13 @@
 
 /**
  * @addtogroup platfIno
- * 
- * This class provides a simple API for connecting and controlling motors.
- * Each motor is assigned to a Tle94112 which acts as output driver. Calls to
- * Tle94112Motor instances are mapped to calls to Tle94112. Therefore, this
- * class does not bring new features, it does only provide further abstraction.
- *
  * @{
  */
 
-class Tle94112MotorIno: public Tle94112Motor
-{
-};
-/** @} */
+
+/**
+ * @}
+ */
 
 #endif /** TLE94112_FRAMEWORK **/
 #endif /** TLE94112MOTOR_PLATF_INO_HPP_ **/
