+++ src/framework/wiced-43xxx/apps/basicTest/basicTest.cpp
@@ -13,6 +13,7 @@
  *
  */
 
+#include "../../../../config/tle94112-conf.hpp"
 #include "basicTest.hpp"
 
 #if (TLE94112_FRAMEWORK == TLE94112_FRMWK_WICED)
