+++ src/framework/wiced-43xxx/apps/control2Motors/control2Motors.mk
@@ -4,6 +4,6 @@ $(NAME)_SOURCES    := control2Motors.cpp
 
 $(NAME)_INCLUDES += ./
 
-$(NAME)_COMPONENTS := drivers/motor-control/multi-half-bridge
+$(NAME)_COMPONENTS := drivers/motor-control/TLE94112
 
 VALID_PLATFORMS	:= CYW943907AEVAL1F
