+++ src/framework/wiced-43xxx/conf_lib.bat
@@ -1,5 +1,5 @@
 @ECHO OFF
-ECHO Setting up apps/ifx/multi-half-bridge hard link...
-mklink /J "../../../../../../../apps/ifx/multi-half-bridge" "apps\"
+ECHO Setting up apps/ifx/TLE94112 hard link...
+mklink /J "../../../../../../../apps/ifx/TLE94112" "apps\"
 ECHO Library configured successfully.
 pause
