+++ src/framework/wiced-43xxx/conf_lib.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
 
-echo "Setting up apps/ifx/multi-half-bridge sym link..."
-ln -s -r app ../../../../../../../apps/ifx/multi-half-bridge
+echo "Setting up apps/ifx/TLE94112 sym link..."
+ln -s -r app ../../../../../../../apps/ifx/TLE94112
 echo "Library configured successfully.
