+++ src/framework/wiced-43xxx/pal/TLE94112-pal-wiced.hpp
@@ -101,7 +101,9 @@ class TLE94112Wiced: virtual public Tle94112
 
 };
 
-/** @} */
+/**
+ * @}
+ */
 
 #endif /** TLE94112_FRAMEWORK **/
 #endif /** TLE94112_WICED_HPP_ **/
