+++ src/framework/wiced-43xxx/wrapper/TLE94112-platf-wiced.cpp
@@ -4,7 +4,6 @@
  * \author      Infineon Technologies AG
  * \copyright   2019 Infineon Technologies AG
  * \version     2.0.0
- * \ref         tle94112wiced
  *
  * SPDX-License-Identifier: MIT
  *
@@ -14,10 +13,6 @@
 
 #if (TLE94112_FRAMEWORK == TLE94112_FRMWK_WICED)
 
-/**
- * @brief Construct a new Tle94112Wiced::Tle94112 Wiced object
- * with default pin assignment
- */
 Tle94112Wiced::Tle94112Wiced(void):Tle94112()
 {
 	Tle94112::en = new GPIOWiced(TLE94112_PIN_EN, OUTPUT_PUSH_PULL, GPIOWiced::POSITIVE );
@@ -26,25 +21,17 @@ Tle94112Wiced::Tle94112Wiced(void):Tle94112()
 	Tle94112::sBus = new SPICWiced();
 }
 
-
-/**
- * @brief constructor with individual pin assignment
- *
- * @param csPin  pin number of the CS pin
- */
-Tle94112Wiced::Tle94112Wiced(wiced_gpio_t csPin):Tle94112()
+Tle94112Wiced::Tle94112Wiced(void* bus, uint8_t csPin):Tle94112()
 {
-	Tle94112::en = new GPIOWiced(TLE94112_PIN_EN, OUTPUT_PUSH_PULL, GPIOWiced::POSITIVE );
-	Tle94112::cs = new GPIOWiced(csPin, OUTPUT_PUSH_PULL, GPIOWiced::POSITIVE );
-	Tle94112::timer = new TimerWiced();
-	Tle94112::sBus = new SPICWiced();
 }
 
-/**
- * @brief constructor with individual pin assignment
- * 
- */
+
 void Tle94112Wiced::begin(void)
+{
+	begin(TLE94112_PIN_CS1);
+}
+
+void Tle94112Wiced::begin(uint8_t csPin)
 {
 	mEnabled = false;
 	Tle94112::sBus->init();
@@ -57,10 +44,6 @@ void Tle94112Wiced::begin(void)
 	init();
 }
 
-/**
- * @brief deactivates all outputs and disables the TLE94112
- * 
- */
 void Tle94112Wiced::end(void)
 {
 	mEnabled = false;
