+++ src/framework/wiced-43xxx/wrapper/TLE94112-platf-wiced.hpp
@@ -4,8 +4,7 @@
  * @date        Mai 2020
  * @version     2.0.0
  * @copyright   Copyright (c) 2019-2020 Infineon Technologies AG
- * @ref         tle94112wiced
- * 
+ *
  * SPDX-License-Identifier: MIT
  */
 
@@ -19,7 +18,7 @@
 #include <wiced_platform.h>
 
 /**
- * @addtogroup platfWiced
+ * @addtogroup wicedHW
  * @{
  */
 
@@ -44,12 +43,32 @@ class Tle94112Wiced: virtual public TLE94112Wiced
 {
 	public:
 
+		//! \brief standard constructor with default pin assignment
 		Tle94112Wiced(void);
-		Tle94112Wiced(wiced_gpio_t csPin);
-		void begin();
+
+		/*! \brief constructor with individual pin assignment
+		 *
+		 * \param bus a void pointer to the object representing the SPI class
+		 * \param csPin  pin number of the CS pin
+		 */
+		Tle94112Wiced(void* bus, uint8_t csPin);
+
+		//! \brief enables and initializes the TLE94112
+		void begin(void);
+
+		/*! \brief constructor with individual pin assignment
+		 *
+		 * \param csPin  pin number of the CS pin
+		 */
+		void begin(uint8_t csPin);
+
+		//! \brief deactivates all outputs and disables the TLE94112
 		void end(void);
+
 };
-/** @} */
+/**
+ * @}
+ */
 
 #endif /** TLE94112_FRAMEWORK **/
 #endif /** TLE94112_PLATF_WICED_HPP_ **/
