+++ src/framework/wiced-43xxx/wrapper/TLE94112Motor-platf-wiced.hpp
@@ -4,8 +4,7 @@
  * @date        Mai 2020
  * @version     2.0.0
  * @copyright   Copyright (c) 2019-2020 Infineon Technologies AG
- * @ref         tle94112wiced
- * 
+ *
  * SPDX-License-Identifier: MIT
  */
 
@@ -19,21 +18,16 @@
 #include "../../../corelib/TLE94112Motor.hpp"
 
 /**
- * @addtogroup platfWiced
- *
- * This class provides a simple API for connecting and controlling motors.
- * Each motor is assigned to a Tle94112 which acts as output driver. Calls to
- * Tle94112Motor instances are mapped to calls to Tle94112. Therefore, this
- * class does not bring new features, it does only provide further abstraction.
- *
+ * @addtogroup wicedHW
  * @{
  */
 
 class Tle94112MotorWiced: virtual public Tle94112Motor
 {
 };
-
-/** @} */
+/**
+ * @}
+ */
 
 #endif /** TLE94112_FRAMEWORK **/
 #endif /** TLE94112MOTOR_PLATF_WICED_HPP_ **/
